(#99) move get solana transaction to controller function
