+++ src/kong_backend/src/solana/get_solana_address.rs
@@ -1,6 +1,6 @@
 use ic_cdk::query;
 
-use crate::stable_memory::{get_cached_solana_address, get_solana_transaction};
+use crate::stable_memory::get_cached_solana_address;
 
 /// Get the cached Solana address for this canister
 /// This is a fast query method that returns the cached address
@@ -15,14 +15,3 @@ pub fn get_solana_address() -> Result<String, String> {
     }
 }
 
-/// DEBUG: Get transaction data for debugging
-#[query]
-pub fn debug_get_transaction(signature: String) -> Result<String, String> {
-    match get_solana_transaction(signature) {
-        Some(tx) => {
-            Ok(format!("Transaction found: status={}, metadata={:?}, timestamp={}", 
-                tx.status, tx.metadata, tx.timestamp))
-        },
-        None => Err("Transaction not found".to_string())
-    }
-}
